Cached libraryType api as that is not going to change in a viewing session. Moved some components around to tighten bundles.
