+++ UI/Web/src/app/library-card/library-card.component.html
