+++ UI/Web/src/app/series-card/series-card.component.html
