+++ UI/Web/src/app/shared/shared.module.ts
@@ -3,14 +3,12 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 import { CardItemComponent } from './card-item/card-item.component';
 import { NgbCollapseModule, NgbDropdownModule, NgbPaginationModule, NgbProgressbarModule, NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';
-import { LibraryCardComponent } from './library-card/library-card.component';
-import { SeriesCardComponent } from './series-card/series-card.component';
 import { CardDetailsModalComponent } from './_modals/card-details-modal/card-details-modal.component';
 import { ConfirmDialogComponent } from './confirm-dialog/confirm-dialog.component';
 import { SafeHtmlPipe } from './safe-html.pipe';
 import { LazyLoadImageModule } from 'ng-lazyload-image';
 import { CardActionablesComponent } from './card-item/card-actionables/card-actionables.component';
-import { RegisterMemberComponent } from './register-member/register-member.component';
+import { RegisterMemberComponent } from '../register-member/register-member.component';
 import { ReadMoreComponent } from './read-more/read-more.component';
 import { RouterModule } from '@angular/router';
 import { DrawerComponent } from './drawer/drawer.component';
@@ -24,8 +22,6 @@ import { A11yClickDirective } from './a11y-click.directive';
   declarations: [
     RegisterMemberComponent,
     CardItemComponent,
-    LibraryCardComponent,
-    SeriesCardComponent,
     CardDetailsModalComponent,
     ConfirmDialogComponent,
     SafeHtmlPipe,
@@ -49,10 +45,8 @@ import { A11yClickDirective } from './a11y-click.directive';
     NgbPaginationModule // CardDetailLayoutComponent
   ],
   exports: [
-    RegisterMemberComponent, // TODO: Move this out and put in normal app
+    RegisterMemberComponent,
     CardItemComponent,
-    LibraryCardComponent, // TODO: Move this out and put in normal app
-    SeriesCardComponent, // TODO: Move this out and put in normal app
     SafeHtmlPipe,
     CardActionablesComponent,
     ReadMoreComponent,
